# Fix `_unsafe_trunc` to reduce the likelihood of arbitrary values