This fixes (mitigates) #288.