@kimikage 2024-04-30
We are having problems testing ColorTypes.jl. So, I am going to backport this to v0.8.
However, since v0.8.4, the files have been rewritten a lot. Any thoughts?