@kimikage 2024-04-30
In any case, this PR is a bug fix, so I will merge it first.
Theoretically, there should be no major performance regression, and even if there is, it should be tackled in another PR.